Put underscore replacements in separate module
