+++ src/cursor.js
@@ -1,36 +1,13 @@
-import isEqual from 'lodash.isequal';
 import getStateByPath from './get_state_by_path';
 import putOnChan from './put_on_chan';
+import { hasMethod, isEqual, partial } from './underscore_ish';
 
 function equals(valueA, cursorB) {
   if (!hasMethod(cursorB, 'deref')) {
     // `cursorB` not a cursor so definitely not equal values
     return false;
   }
-  return valuesEqual(valueA, cursorB.deref());
-}
-
-function valuesEqual(valueA, valueB) {
-  if (valueA === valueB) {
-    return true;
-  }
-
-  // try an equals method
-  if (hasMethod(valueA, 'equals') && hasMethod(valueB, 'equals')) {
-    return valueA.equals(valueB);
-  }
-
-  // try result of valueOf method, only if it returns a distinct value
-  if (hasMethod(valueA, 'valueOf') && hasMethod(valueB, 'valueOf') && valueA.valueOf() !== valueA) {
-    return (valueA.valueOf() === valueB.valueOf());
-  }
-
-  // let lodash figure it out
-  return isEqual(valueA, valueB);
-}
-
-function hasMethod(o, name) {
-  return o != null && typeof o === 'object' && typeof o[name] === 'function';
+  return isEqual(valueA, cursorB.deref());
 }
 
 function deref(value) {
@@ -63,7 +40,7 @@ function fetch(ch, path) {
 }
 
 function update(ch, change) {
-  putOnChan(ch, change);
+  putOnChan(ch, Object.freeze(change));
 }
 
 // refining
@@ -78,28 +55,28 @@ function refine(value, updateCh, oldPath, newPath) {
 const cursor = function (value, path, updateCh) {
 
   let o = {
-    deref:    deref.bind(this, value),
+    deref:    partial(deref, value),
     path:     path,
 
-    replace:  replace.bind(this, updateCh, path),
-    remove:   remove.bind(this, updateCh, path, value),
+    replace:  partial(replace, updateCh, path),
+    remove:   partial(remove, updateCh, path, value),
 
-    refine:   refine.bind(this, value, updateCh, path),
+    refine:   partial(refine, value, updateCh, path),
 
-    persist:  persist.bind(this, updateCh, path),
-    fetch:    fetch.bind(this, updateCh, path),
+    persist:  partial(persist, updateCh, path),
+    fetch:    partial(fetch, updateCh, path),
 
-    equals:   equals.bind(this, value)
+    equals:   partial(equals, value)
   };
 
   if (Array.isArray(value)) {
     // array specific operations
-    o.add = add.bind(this, updateCh, path);
+    o.add = partial(add, updateCh, path);
   }
 
   if (hasMethod(value, 'valueOf')) {
     // object specific operations
-    o.set = set.bind(this, updateCh, path);
+    o.set = partial(set, updateCh, path);
   }
 
   return o;
